requirements.txt
Babel==2.9.1
celery==5.2.3
django-bootstrap5==21.3
django-celery-results==2.2.0
django-filter==21.1
django-import-export==2.7.1
django-phonenumber-field==6.1.0
Django==3.2.12
get-docker-secret==1.0.2
gunicorn==20.1.0
phonenumbers==8.12.45
psycopg2-binary==2.9.3
python-dateutil==2.8.2
requests==2.28.2
Unidecode==1.3.4
wait-for-it==2.2.1
whitenoise==6.0.0
